feat: history sort 기능 추가
